+++ packages/picasso/src/Icon/Education16.tsx
@@ -6,6 +6,7 @@ import { StandardProps } from '@toptal/picasso-shared'
 import kebabToCamelCase from '../utils/kebab-to-camel-case'
 import styles from './styles'
 const BASE_SIZE = 16
+
 type ScaleType = 1 | 2 | 3 | 4
 export interface Props extends StandardProps {
   scale?: ScaleType
@@ -37,6 +38,7 @@ const SvgEducation16 = forwardRef(function SvgEducation16(
     minHeight: `${scaledSize}px`,
     ...style
   }
+
   return (
     <svg
       viewBox='0 0 16 16'
@@ -48,5 +50,6 @@ const SvgEducation16 = forwardRef(function SvgEducation16(
     </svg>
   )
 })
+
 SvgEducation16.displayName = 'SvgEducation16'
 export default withStyles(styles)(SvgEducation16)