+++ packages/picasso/src/Modal/Modal.tsx
@@ -26,6 +26,7 @@ import ModalActions from '../ModalActions'
 import styles from './styles'
 
 type ContainerValue = HTMLElement | (() => HTMLElement)
+type Alignment = 'top' | 'centered'
 
 export interface Props extends BaseProps, HTMLAttributes<HTMLDivElement> {
   /** Content of Modal component */
@@ -44,6 +45,8 @@ export interface Props extends BaseProps, HTMLAttributes<HTMLDivElement> {
   container?: ContainerValue
   /** If `true`, the backdrop is not rendered */
   hideBackdrop?: boolean
+  /** Position of the modal relative to the browser's viewport */
+  align?: Alignment
   transitionDuration?: number
   paperProps?: PaperProps
 }
@@ -82,14 +85,14 @@ const isFocusInsideModal = (modalNode: Element) => {
 }
 
 const isFocusInsideTooltip = () => {
-  const tooltipContainer = document.querySelector(tooltipContainerString)
+  const tooltipContainers = document.querySelectorAll(tooltipContainerString)
 
-  if (!tooltipContainer) {
+  if (tooltipContainers.length === 0) {
     return false
   }
 
-  const tooltipContainsFocusedElement = tooltipContainer.contains(
-    document.activeElement
+  const tooltipContainsFocusedElement = Array.from(tooltipContainers).some(
+    container => container.contains(document.activeElement)
   )
 
   if (tooltipContainsFocusedElement) {
@@ -120,6 +123,7 @@ export const Modal = forwardRef<HTMLElement, Props>(function Modal(props, ref) {
     hideBackdrop,
     transitionDuration,
     paperProps,
+    align,
     ...rest
   } = props
   const classes = useStyles(props)
@@ -160,6 +164,7 @@ export const Modal = forwardRef<HTMLElement, Props>(function Modal(props, ref) {
   }, [open, rootRef])
 
   const bodyOverflow = useRef<string>(document.body.style.overflow)
+
   useEffect(() => {
     const resetBodyOverflow = () => {
       document.body.style.overflow = bodyOverflow.current
@@ -194,7 +199,9 @@ export const Modal = forwardRef<HTMLElement, Props>(function Modal(props, ref) {
       classes={{
         root: classes.root,
         container: classes.container,
-        paper: cx(classes.paper, classes[size!])
+        paper: cx(classes.paper, classes[size!], {
+          [classes.topAlignedDialog]: align === 'top'
+        })
       }}
       className={className}
       style={style}
@@ -223,7 +230,8 @@ export const Modal = forwardRef<HTMLElement, Props>(function Modal(props, ref) {
 Modal.defaultProps = {
   hideBackdrop: false,
   size: 'medium',
-  transitionDuration: 300
+  transitionDuration: 300,
+  align: 'centered'
 }
 
 Modal.displayName = 'Modal'