+++ packages/picasso/src/Modal/story/Default.example.tsx
@@ -14,13 +14,19 @@ const STATES = [
   }
 ]
 
-const ModalDialog = ({ modalId, hideModal }) => {
+const ModalDialog = ({
+  modalId,
+  hideModal
+}: {
+  modalId: string
+  hideModal: (modalId: string) => void
+}) => {
   const [isLoading, setLoading] = useState(false)
-  const [datepickerValue, setDatepickerValue] = useState()
+  const [datepickerValue, setDatepickerValue] = useState<Date>()
 
   return (
     <Modal
-      container={() => document.getElementById('modal-container')}
+      container={() => document.getElementById('modal-container')!}
       onBackdropClick={() => console.log('Clicked backdrop..')}
       onClose={() => hideModal(modalId)}
       onOpen={() => console.log('onOpen()')}
@@ -46,7 +52,7 @@ const ModalDialog = ({ modalId, hideModal }) => {
               /* eslint-disable-next-line no-console */
               console.log('selected date is: ', date)
 
-              setDatepickerValue(date)
+              setDatepickerValue(date as Date)
             }}
           />
         </Form.Field>