+++ packages/picasso/src/ScrollMenu/ScrollMenu.tsx
@@ -1,6 +1,5 @@
 import React, {
-  useState,
-  useEffect,
+  useLayoutEffect,
   useRef,
   createRef,
   FunctionComponent
@@ -35,6 +34,10 @@ const getMoveDirection = (
     : Direction.UP
 }
 
+const preventDefault = (
+  event: React.MouseEvent<HTMLUListElement, MouseEvent>
+) => event.preventDefault()
+
 const ScrollMenu: FunctionComponent<Props> = ({
   selectedIndex,
   classes,
@@ -43,7 +46,7 @@ const ScrollMenu: FunctionComponent<Props> = ({
 }) => {
   const menuRef = useRef<HTMLDivElement | null>(null)
   const firstItemRef = createRef<HTMLElement>()
-  const [prevSelectedIndex, setPrevSelectedIndex] = useState(selectedIndex)
+  const prevSelectedIndex = useRef(selectedIndex)
 
   const renderChildren = React.Children.map(children, (child, index) => {
     if (index === 0 && child) {
@@ -53,7 +56,7 @@ const ScrollMenu: FunctionComponent<Props> = ({
     return child
   })
 
-  useEffect(() => {
+  useLayoutEffect(() => {
     if (!menuRef.current || !firstItemRef.current) {
       return
     }
@@ -76,7 +79,7 @@ const ScrollMenu: FunctionComponent<Props> = ({
     if (!isHighlightedItemInScrollView) {
       const moveDirection = getMoveDirection(
         selectedIndex,
-        prevSelectedIndex,
+        prevSelectedIndex.current,
         bottomVisibleItem
       )
       let scrollTop = 0
@@ -90,11 +93,11 @@ const ScrollMenu: FunctionComponent<Props> = ({
       menuRef.current.scrollTop = scrollTop
     }
 
-    setPrevSelectedIndex(selectedIndex)
+    prevSelectedIndex.current = selectedIndex
   }, [firstItemRef, selectedIndex, prevSelectedIndex])
 
   return (
-    <Menu className={classes.menu} style={style}>
+    <Menu className={classes.menu} style={style} onMouseDown={preventDefault}>
       <div ref={menuRef} className={classes.scrollView}>
         {renderChildren}
       </div>