+++ packages/picasso-forms/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.20.5](https://github.com/toptal/picasso/compare/@toptal/picasso-forms@0.20.4...@toptal/picasso-forms@0.20.5) (2020-09-09)
+
+**Note:** Version bump only for package @toptal/picasso-forms
+
+
+
+
+
+## [0.20.4](https://github.com/toptal/picasso/compare/@toptal/picasso-forms@0.20.3...@toptal/picasso-forms@0.20.4) (2020-08-31)
+
+**Note:** Version bump only for package @toptal/picasso-forms
+
+
+
+
+
+## [0.20.3](https://github.com/toptal/picasso/compare/@toptal/picasso-forms@0.20.2...@toptal/picasso-forms@0.20.3) (2020-08-31)
+
+
+### Bug Fixes
+
+* add support for complex field validation ([#1513](https://github.com/toptal/picasso/issues/1513)) ([ddf80e7](https://github.com/toptal/picasso/commit/ddf80e773f4fae64e420e6931740340e1894fe58))
+
+
+
+
+
+## [0.20.2](https://github.com/toptal/picasso/compare/@toptal/picasso-forms@0.20.1...@toptal/picasso-forms@0.20.2) (2020-08-27)
+
+
+### Bug Fixes
+
+* **Form:** improve form errors scrolling ([#1512](https://github.com/toptal/picasso/issues/1512)) ([16f9dd6](https://github.com/toptal/picasso/commit/16f9dd62fad2979d129fc53993da3e7a66d622cb))
+
+
+
+
+
 ## [0.20.1](https://github.com/toptal/picasso/compare/@toptal/picasso-forms@0.20.0...@toptal/picasso-forms@0.20.1) (2020-08-13)